refactor contracts diamond
