+++ packages/contracts-diamond/src/Diamond.ts
@@ -1,16 +1,6 @@
-import { Clients, getDeployContractsParams, getDeployFunctionData } from "@owlprotocol/contracts-create2factory";
-import {
-    Abi,
-    Address,
-    Hex,
-    encodeDeployData,
-    encodeFunctionData,
-    toFunctionSelector,
-    zeroAddress,
-    zeroHash,
-} from "viem";
-import { Diamond, DiamondInit, DiamondInitMulti, ERC721MintableAutoIdBaseURIFacetInit } from "./artifacts/index.js";
-import { ERC721PresetDiamondSpec } from "./diamondSpec.js";
+import { Clients } from "@owlprotocol/contracts-create2factory";
+import { Abi, Address, Hex, encodeDeployData, encodeFunctionData, toFunctionSelector, zeroAddress } from "viem";
+import { Diamond, DiamondInitMulti } from "./artifacts/index.js";
 
 export function getAbiFunctionSelectors(abi: Abi): Hex[] {
     return abi
@@ -39,7 +29,7 @@ export enum FacetCutAction {
  * @param args
  * @returns
  */
-export function valodateDiamondCutUpgrades(
+export function validateDiamondCutUpgrades(
     facets: { facetAddress: Address; functionSelectors: Hex[] }[],
 ): { facetAddress: Address; functionSelectors: Hex[]; action: FacetCutAction }[] {
     //TODO: Add logic above, for now just return all
@@ -145,7 +135,7 @@ export interface GetDiamondDeployDataArg {
 export function getDiamondDeployData(admin: Address, deployArgs: GetDiamondDeployDataArg) {
     const diamondInitData = getDiamondMultiInitData(deployArgs.initializers);
 
-    const diamondCutUpgrades = valodateDiamondCutUpgrades(deployArgs.facets);
+    const diamondCutUpgrades = validateDiamondCutUpgrades(deployArgs.facets);
 
     return {
         diamondCutUpgrades,
@@ -155,81 +145,5 @@ export function getDiamondDeployData(admin: Address, deployArgs: GetDiamondDeplo
             bytecode: Diamond.bytecode,
             args: [admin, diamondCutUpgrades, diamondInitData.to, diamondInitData.data],
         }),
-    };
-}
-
-//TODO: Use zod codegen for this
-interface DiamondInitParams {
-    admin: Address;
-}
-interface ERC721InitParams {
-    contractUri: string;
-    name: string;
-    symbol: string;
-    baseUri: string;
-    royaltyReceiver: Address;
-    feeNumerator: bigint;
-}
-export function getERC721DiamondDeployData(initParams: ERC721InitParams & DiamondInitParams) {
-    const { admin, contractUri, name, symbol, baseUri, royaltyReceiver, feeNumerator } = initParams;
-    const facets = Object.values(ERC721PresetDiamondSpec.facets).map((f) => {
-        return {
-            facetAddress: f.implementation,
-            functionSelectors: getAbiFunctionSelectors(f.abi),
-        };
-    });
-    const initializers = [
-        {
-            abi: DiamondInit.abi,
-            address: DiamondInit.implementation,
-            functionName: "initialize",
-            args: [],
-        },
-        {
-            abi: ERC721MintableAutoIdBaseURIFacetInit.abi,
-            address: ERC721MintableAutoIdBaseURIFacetInit.implementation,
-            functionName: "initialize",
-            args: [contractUri, name, symbol, baseUri, royaltyReceiver, feeNumerator],
-        },
-    ];
-    const { deployData, diamondCutUpgrades, diamondInitData } = getDiamondDeployData(admin, { facets, initializers });
-    return {
-        facets,
-        initializers,
-        diamondCutUpgrades,
-        diamondInitData,
-        deployData,
-    };
-}
-
-export async function getERC721ImplementationDeployParams(clients: Pick<Clients, "publicClient">) {
-    const artifacts = [
-        ...Object.values(ERC721PresetDiamondSpec.facets),
-        ...Object.values(ERC721PresetDiamondSpec.initializers),
-    ];
-    //If implementation not deployed, we add it to  `Create2Factory.deployDeterministic` call
-    // - Diamond (core)
-    // - Diamond facets (selector-specific implementations)
-    // - Diamond inits (delegate call post-cut)
-    // `getDeployContractsParams` returns only contracts required for deployment
-    const deployParams = await getDeployContractsParams(
-        clients,
-        zeroAddress,
-        Object.values(artifacts).map((a) => {
-            return { salt: zeroHash, bytecode: a.bytecode, initData: "0x" };
-        }),
-    );
-
-    // If deployed addressesNew.length > 0, we compute the call data, otherwise undefined
-    let deployTransaction: { to: Address; data: Hex } | undefined;
-    if (deployParams.addressesNew.length > 0) {
-        deployTransaction = getDeployFunctionData(zeroAddress, deployParams.contractsNew);
-    }
-
-    return {
-        deployParams: deployParams.contractsNew,
-        deployTransaction,
-        addresses: deployParams.addressesAll,
-        addressesNew: deployParams.addressesNew,
     };
 }
